refactoring: security variable was moved, fix security
