+++ erp-core/src/main/java/ru/aora/erp/config/SecurityConfig.java
@@ -15,17 +15,18 @@ import ru.aora.erp.service.UserService;
 
 import java.util.Map;
 
-import static ru.aora.erp.component.CoreModuleIdentifier.DASHBOARD_MAPPING;
-import static ru.aora.erp.component.CoreModuleIdentifier.LOGIN_MAPPING;
-import static ru.aora.erp.component.CoreModuleIdentifier.LOGOUT_MAPPING;
+import static ru.aora.erp.controller.DashboardController.DASHBOARD_MAPPING;
+import static ru.aora.erp.controller.SecurityController.INCLUDE_ROOT_MAPPING;
+import static ru.aora.erp.controller.SecurityController.LOGIN_MAPPING;
+import static ru.aora.erp.controller.SecurityController.LOGOUT_MAPPING;
 
 
 @Configuration
 @EnableWebSecurity
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
-    private static final String ALL_CSS_MAPPING = "/css/**";
     private static final String ALL_JS_MAPPING = "/js/**";
+    private static final String ALL_CSS_MAPPING = "/css/**";
     private static final String ALL_ICONS_MAPPING = "/icons/**";
 
     private final UserService userService;
@@ -45,8 +46,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
     @Override
     protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.userDetailsService(userService)
-                .passwordEncoder(passwordEncoder);
+        auth.userDetailsService(userService).passwordEncoder(passwordEncoder);
     }
 
     @Override
@@ -57,22 +57,17 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers(ALL_JS_MAPPING).permitAll()
                 .antMatchers(ALL_ICONS_MAPPING).permitAll()
 //                .antMatchers(INCLUDE_ROOT_MAPPING).hasAnyAuthority(UserRole.USER.getAuthority(),UserRole.ADMIN.getAuthority())
+                .antMatchers(INCLUDE_ROOT_MAPPING).permitAll()
 //                .antMatchers(DASHBOARD_MAPPING).hasAnyAuthority(UserRole.USER.getAuthority(),UserRole.ADMIN.getAuthority())
-                .antMatchers("/**").permitAll()
-                .anyRequest().authenticated()
-
-                .and()
-                .formLogin()
+//                .antMatchers(INCLUDE_ROOT_MAPPING).authenticated()
+                .anyRequest().authenticated();
+        http.formLogin()
                 .loginPage(LOGIN_MAPPING)
                 .defaultSuccessUrl(DASHBOARD_MAPPING)
                 .successForwardUrl(DASHBOARD_MAPPING)
-                .permitAll()
-
-                .and()
-                .logout()
-                .logoutRequestMatcher(
-                        new AntPathRequestMatcher(LOGOUT_MAPPING)
-                )
+                .permitAll();
+        http.logout()
+                .logoutRequestMatcher(new AntPathRequestMatcher(LOGOUT_MAPPING))
                 .logoutSuccessUrl(LOGIN_MAPPING);
 //        defineAuthoritiesMapping(http);
     }
