+++ erp-core/src/main/java/ru/aora/erp/controller/DashboardController.java
@@ -7,12 +7,13 @@ import ru.aora.erp.service.SidebarModulesIdentifiersService;
 
 import java.util.Map;
 
-import static ru.aora.erp.component.CoreModuleIdentifier.DASHBOARD_MAPPING;
-import static ru.aora.erp.component.CoreModuleIdentifier.ROOT_MAPPING;
+import static ru.aora.erp.controller.SecurityController.ROOT_MAPPING;
 import static ru.aora.erp.utils.mvc.MvcUtils.redirectTo;
 
 @Controller
 public final class DashboardController {
+
+    public static final String DASHBOARD_MAPPING = "/dashboard";
 
     private static final String DASHBOARD_TEMPLATE = "dashboard";
     private static final String UI_CHANE_NODE_MODEL = "uiChaneNodeModel";
