+++ erp-core/src/main/java/ru/aora/erp/controller/KsController.java
@@ -19,7 +19,7 @@ import java.util.Map;
 
 @Controller
 @RequestMapping("/ks")
-public class KsController {
+public final class KsController {
 
     private static final String CONTROLLER_MAPPING = "kss";
     private static final String DTO_MODEL = "ksDto";
